requirements.txt
cycler==0.11.0
fonttools==4.27.1
kiwisolver==1.4.4
matplotlib==3.5.2
networkx==2.8.7
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.3
Pillow==8.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.1
six==1.16.0
tk==0.1.0
